Unify `parseList` implementations and move all expression parsing to `ParserBase`.
